kernel: check if two files are identical, if so, skip copying

+++ src/kernel/mod.rs
@@ -1,5 +1,6 @@
 use anyhow::{bail, Result};
-use std::{fmt::Display, fs, path::Path};
+use sha2::{Sha256, Digest};
+use std::{fmt::Display, fs, io, path::Path};
 
 use crate::fl;
 
@@ -15,9 +16,25 @@ pub trait Kernel: Display + Clone + PartialEq {
     fn install_and_make_config(&self, force_write: bool) -> Result<()>;
 }
 
+// Check if two files are identical
+fn same_files<P: AsRef<Path>>(file1: P, file2: P) -> Result<bool> {
+    let mut hasher1 = Sha256::new();
+    io::copy(&mut fs::File::open(&file1)?, &mut hasher1)?;
+    let hash1 = hasher1.finalize();
+
+    let mut hasher2 = Sha256::new();
+    io::copy(&mut fs::File::open(&file2)?, &mut hasher2)?;
+    let hash2 = hasher2.finalize();
+
+    Ok(hash1 == hash2)
+}
+
 // Make sure the copy is complete, otherwise possible ENOSPC (No space left on device)
 pub fn safe_copy<P: AsRef<Path>>(src: P, dest: P) -> Result<()> {
-    if fs::metadata(&src)?.len() != fs::copy(&src, &dest)? {
+    if dest.as_ref().exists()
+        && !same_files(&src, &dest)?
+        && fs::metadata(&src)?.len() != fs::copy(&src, &dest)?
+    {
         // Remove incomplete copy
         fs::remove_file(&dest)?;
         bail!(fl!("no_space"));
